backend/AI/requirements-gpu.txt
--extra-index-url https://download.pytorch.org/whl/cu121

ultralytics==8.3.1
torch==2.2.1+cu121 # CUDA 12.1 빌드 (pip index URL 필요, 아래 참고)
torchvision==0.17.1+cu121
pillow>=10.0
opencv-python-headless>=4.9
